requirements.txt
asgiref==3.3.1
da-vinci==0.3.0
Django==3.1.7
django-cors-headers==3.7.0
django-money==1.3.1       
django-thumbnails==0.4.0
djangorestframework==3.12.4
Pillow==8.2.0
py-moneyed==0.8.0
pytz==2021.1
redis==3.5.3
shortuuid==1.0.1
sqlparse==0.4.1